+++ find_sys_messge.js
@@ -17,11 +17,11 @@ async function main() {
         process.exit(1)
     }
 
-    let store = 'mongodb://localhost/trop_diary'
+    let store = 'mongodb://localhost/trop_diary_api'
     let client = await MongoClient.connect(store, {
         useNewUrlParser: true
     })
-    let db = client.db('trop_diary')
+    let db = client.db('trop_diary_api')
     let doc = db.collection('system')
     let log = await doc.findOne({
         _id: uuid.from(process.argv[2])
@@ -31,22 +31,31 @@ async function main() {
     } else {
         console.log(_title('level'), _format_level(log.level))
         console.log(_title('created'), _format_date(log.created))
-        console.log(_title('path'), log.message.req_path)
-        console.log(_title('req_method'), log.message.req_method)
-        console.log(_title('res_status'), log.message.res_status)
-        console.log(_title('req_path'), log.message.req_path)
-        console.log(_title('req_query'), log.message.req_query)
-        console.log(_title('req_body'), log.message.req_body)
+        console.log(_title('req_method'), log.req_method)
+        console.log(_title('res_status'), log.res_status)
         console.log()
-        console.log('===ERROR_BEGIN===')
-        console.log(log.message.error)
-        console.log('===ERROR_END===')
+        console.log(_block('req_path'))
+        console.log(log.req_path)
+        console.log(_block('req_query'))
+        console.log(log.req_query)
+        console.log(_block('req_body'))
+        console.log(log.req_body)
+        console.log(_block('front'))
+        console.log(JSON.stringify(log.front, null, 2))
+        console.log(_block('back'))
+        console.log(JSON.stringify(log.back, null, 2))
+        console.log(_block('stack'))
+        console.log(log.stack)
     }
     await client.close()
 }
 
 function _title(title) {
     return title.toUpperCase().padEnd(16)
+}
+
+function _block(title) {
+    return '====' + title.toUpperCase() + '===='
 }
 
 function _format_level(no) {